[Update]feat: ラズパイのオーナー登録/登録解除の機能を実装

+++ src/v1/account/account.service.ts
@@ -7,10 +7,10 @@ export class AccountService {
   //*--------------------------------------------
   //* PostメソッドでUserDBの作成をする場合
   //*--------------------------------------------
-  // 作成済み    => resHttpStatus 200
-  // 作成に成功  => resHttpStatus 201
-  // 作成に失敗  => resHttpStatus 500
-  // Tokenエラー => resHttpStatus 401
+  // 作成済み    => resHttpStatus 200(HttpStatus.OK)
+  // 作成に成功  => resHttpStatus 201(HttpStatus.CREATED)
+  // 作成に失敗  => resHttpStatus 500(HttpStatus.INTERNAL_SERVER_ERROR)
+  // Tokenエラー => resHttpStatus 401(HttpStatus.UNAUTHORIZED)
   async create(AccountDto: AccountDto) {
     // return用Httpステータスコード
     let resHttpStatus: number = undefined;
@@ -20,40 +20,55 @@ export class AccountService {
       .then((decodedToken) => {
         return decodedToken.uid;
       })
-      .catch((error) => {
-        console.log('---------- Error ----------\nverifyIdToken:');
-        console.log(error);
+      .catch((err) => {
+        console.log('Error: auth.verifyIdToken(AccountDto.Token)');
+        console.log(err);
+        // トークンエラーの場合はステータスコード401を代入
+        resHttpStatus = HttpStatus.UNAUTHORIZED;
         return undefined;
       });
-    // FirestoreDatabaseにUsers/uidがあるか存在確認
+
+    //* uidが登録された正規のuidだった場合
     if (uid != undefined) {
+      //* FirestoreDatabaseにUsers/uidがあるか存在確認
       await refFirestoreDbUser
         .doc(uid)
         .get()
-        .then(() => {
-          // 読み取れた場合は作成済みなのでステータスコード200を代入
-          resHttpStatus = HttpStatus.OK;
+        .then((doc) => {
+          console.log(`doc.exists: ${doc.exists}`);
+          if (doc.exists) {
+            // 作成済み の場合はステータスコード200を代入
+            resHttpStatus = HttpStatus.OK;
+          }
+          // 未作成 の場合はステータスコード201を代入
+          else {
+            resHttpStatus = HttpStatus.CREATED;
+          }
         })
-        .catch((error) => {
+        .catch((err) => {
           console.log('Error: refFirestoreDbUser.doc(uid).get()');
-          console.error(error);
+          console.error(err);
           resHttpStatus = HttpStatus.CREATED; // エラーの場合は未作成なのでステータスコード201を代入
         });
     }
+    //*/
 
     //* uidが登録された正規のuid && FirestoreDbにユーザーデータが無かった場合
-    if (uid != undefined && resHttpStatus == 201) {
-      const pushData = { RaspPiSerialNumber: [], UserName: null, imageURL: '' };
+    if (uid != undefined && resHttpStatus == HttpStatus.CREATED) {
+      const pushData = { RaspPiSerialNumber: [], UserName: '', imageURL: '' };
       await refFirestoreDbUser
         .doc(uid)
         .set(pushData)
         .catch((err) => {
           console.log('Error: refFirestoreDbUser.doc(uid).set(pushData)');
           console.log(err);
+          // エラーが発生した場合はステータスコード401を代入
           resHttpStatus = HttpStatus.INTERNAL_SERVER_ERROR;
         });
     }
+    //*/
 
+    //* return
     console.log(`resHttpStatus: ${resHttpStatus}\n`);
     return resHttpStatus;
   }
@@ -61,10 +76,10 @@ export class AccountService {
   //*--------------------------------------------
   //* DELETEメソッドでUserデータを削除する場合
   //*--------------------------------------------
-  // 削除済み    => resHttpStatus 200
-  // 削除に成功  => resHttpStatus 204
-  // 削除に失敗  => resHttpStatus 500
-  // Tokenエラー => resHttpStatus 401
+  // 削除済み    => resHttpStatus 200(HttpStatus.OK)
+  // 削除に成功  => resHttpStatus 204(HttpStatus.NO_CONTENT)
+  // 削除に失敗  => resHttpStatus 500(HttpStatus.INTERNAL_SERVER_ERROR)
+  // Tokenエラー => resHttpStatus 401(HttpStatus.UNAUTHORIZED)
   async remove(AccountDto: AccountDto) {
     // return用Httpステータスコード
     let resHttpStatus: number;
@@ -74,60 +89,55 @@ export class AccountService {
       .then((decodedToken) => {
         return decodedToken.uid;
       })
-      .catch((error) => {
-        console.log('---------- Error ----------\nverifyIdToken:');
-        console.log(error);
+      .catch((err) => {
+        console.log('Error: auth.verifyIdToken(AccountDto.Token)');
+        console.log(err);
+        // トークンエラーの場合はステータスコード401を代入
+        resHttpStatus = HttpStatus.UNAUTHORIZED;
         return undefined;
       });
 
     //* uidが登録された正規のuidだった場合
     if (uid != undefined) {
-      resHttpStatus = HttpStatus.NO_CONTENT;
-      let isFirestoreDbExists: boolean = undefined;
-
       //* FirestoreDatabaseにUsers/uidがあるか存在確認
-      // 結果true/falseをisFirestoreDbExistsに格納
       await refFirestoreDbUser
         .doc(uid)
         .get()
         .then((doc) => {
-          if (doc.exists) isFirestoreDbExists = true;
+          console.log(`doc.exists: ${doc.exists}`);
+          // データが存在 の場合はステータスコード204を代入
+          if (doc.exists) {
+            resHttpStatus = HttpStatus.NO_CONTENT;
+          }
+          // データが無かった の場合はステータスコード200を代入
+          else {
+            resHttpStatus = HttpStatus.OK;
+          }
         })
-        .catch((error) => {
+        .catch((err) => {
           console.log('Error: refFirestoreDbUser.doc(uid).get()');
-          console.error(error);
-          resHttpStatus = HttpStatus.INTERNAL_SERVER_ERROR;
-        })
-        .finally(() => {
-          console.log(`isFirestoreDbExists: ${isFirestoreDbExists}`);
+          console.error(err);
+          // 存在していなかった場合はステータスコード200を代入
+          resHttpStatus = HttpStatus.OK;
         });
-      //*/
-
-      //* FirestoreDbにユーザーデータが有った場合
-      if (
-        isFirestoreDbExists == true &&
-        resHttpStatus != HttpStatus.INTERNAL_SERVER_ERROR
-      ) {
-        await refFirestoreDbUser
-          .doc(uid)
-          .delete()
-          .catch((err) => {
-            console.log('Error: refFirestoreDbUser.doc(uid).delete()');
-            console.log(err);
-            resHttpStatus = HttpStatus.INTERNAL_SERVER_ERROR;
-          });
-      }
-      //* 削除済みだった場合
-      else if (
-        isFirestoreDbExists != true &&
-        resHttpStatus != HttpStatus.INTERNAL_SERVER_ERROR
-      ) {
-        resHttpStatus = HttpStatus.OK;
-      }
-    } else {
-      resHttpStatus = HttpStatus.UNAUTHORIZED;
     }
+    //*/
 
+    //* FirestoreDbにユーザーデータが有った場合
+    if (uid != undefined && resHttpStatus == HttpStatus.NO_CONTENT) {
+      await refFirestoreDbUser
+        .doc(uid)
+        .delete()
+        .catch((err) => {
+          console.log('Error: refFirestoreDbUser.doc(uid).delete()');
+          console.log(err);
+          // エラーが発生した場合はステータスコード401を代入
+          resHttpStatus = HttpStatus.INTERNAL_SERVER_ERROR;
+        });
+    }
+    //*/
+
+    //* return
     console.log(`resHttpStatus: ${resHttpStatus}\n`);
     return resHttpStatus;
   }
